:recycle: refactor: UserException -> UserHandler 로 클래스명 변경
